refactor: consolidate platform code and split monolithic files
Contain platform specifics in seance-app::platform and break up the
two largest files in the workspace.

Platform consolidation:
- Move configure_metal_layer out of seance-render::gpu::state into
  seance-app::platform. The call still runs after the wgpu surface is
  created (the CAMetalLayer is attached as a sublayer at that point),
  just relocated to the app so seance-render can stay platform-agnostic.
- Add platform::configure_event_loop to absorb the macOS ActivationPolicy
  cfg block that lived inline in main(). No-op stubs on non-macOS keep
  cfg attributes out of main.rs.
- Drop objc2 and raw-window-handle from seance-render/Cargo.toml —
  nothing in the crate uses them now that the Metal layer setup moved.

seance-app/src/main.rs (744 -> 35 lines):
- app.rs: App struct, ApplicationHandler impl, per-event handlers,
  helpers (physical_window_padding, option_as_alt_from_config,
  vt_shape_from_config).
- mouse.rs: MouseState + multi-click debouncing.
- reload.rs: hot-reload impls (reload_config, reload_theme,
  on_theme_file_changed, apply_terminal_theme).
- main.rs: entry point and UserEvent only.

seance-render/src/gpu/state.rs (536 -> 376 lines):
- dynamic_buffer.rs: DynamicBuffer (grow-on-demand wgpu buffer).
- atlas_texture.rs: write_atlas_plane / atlas_view / bytes_per_pixel
  helpers shared by the grayscale and color atlas planes.

No behavior change. Tests that pass locally continue to pass; the five
seance-config theme failures in this sandbox are pre-existing (the
vendor/ghostty-themes clone failed with a GitHub 500).
