+++ crates/seance-render/Cargo.toml
@@ -18,7 +18,3 @@ winit = "0.30"
 
 [lints]
 workspace = true
-
-[target.'cfg(target_os = "macos")'.dependencies]
-objc2 = "0.6"
-raw-window-handle = "0.6"
