+++ crates/seance-render/src/gpu/mod.rs
@@ -1,3 +1,5 @@
+mod atlas_texture;
+mod dynamic_buffer;
 mod pipeline;
 mod state;
 pub(crate) mod uniforms;
